refactor(spring-server-all): 将用户标识从account改为 ID
- 修改了多个接口和服务中的用户账号参数，改为使用用户 ID（Long 类型）
- 更新了相关 DTO 和实体类中的字段，从账号改为用户 ID
- 调整了部分方法名称，以适应新的参数类型
- 移除了不必要的账号相关代码，简化了部分逻辑

+++ spring-server/business/user-relationship-service/src/main/java/com/czy/user/mapper/mysql/relation/UserFriendMapper.java
@@ -35,10 +35,10 @@ public interface UserFriendMapper {
     List<UserViewEntity> getUserFriendsViewByAccount(String userAccount);
 
     // 通过userId查询用户好友的全部好友及其view信息
-    List<UserViewEntity> getUserFriendsViewByUserId(int userId);
+    List<UserViewEntity> getUserFriendsViewByUserId(Long userId);
 
     // 查询用户的好友数量
-    int getUserFriendsCount(int userId);
+    int getUserFriendsCount(Long userId);
 
     // 查询两个用户是否是好友
     int isFriend(@Param("userId") Long userId, @Param("friendId") Long friendId);
